// backend/src/common/enums.ts
export enum Perfil {
  ADMIN = 'ADMIN',
  GESTOR = 'GESTOR',
  COLABORADOR = 'COLABORADOR',
}

export enum Status {
  EM_ATRIBUICAO = 'em_atribuicao',
  EM_FILA = 'em_fila',
  DESENVOLVIMENTO = 'desenvolvimento',
  APROVACAO = 'aprovacao',
  DOCUMENTACAO = 'documentacao',
  ENTREGUE = 'entregue',
}